Sync with upstream (#74)
* v0.4.12

* test: window update stream errors count towards local max (#859)

test: more local reset counts

* ci: pin indexmap and tracing for msrv job (#866)

* fix: Unparent connection span (#868)

The `Connection` struct contains a `tracing::Span` tracking its
lifetime. This can be useful. The problem is that this span has as its
parent the span within which the connection is created, meaning that it
keeps that span alive (not closed) while it is alive. In practice that
means, for example, instrumented functions that create connections and
store those connections for reuse will not have their spans closed and
exported when they end.

This commit sets the `Connection`'s span's parent to `None` and gives it
a "follows-from" relation to the span in which it was created instead.
This shows the causal relationship between them, without tying the
creating span's lifetime to the connection's lifetime.

---------

Co-authored-by: Sean McArthur <sean@seanmonstar.com>
Co-authored-by: Jesse <github@jessebakker.com>